refactor(tools): extract response helpers from handlers.go into response_helpers.go
Move okText, okJSON, okWithWarning, errorf, blockGuard, loadState,
validateTaskDependencies, and publishEvent into a dedicated file so
handlers.go contains only the 26 *Handler functions.
